Split DcsOutputAndColorBinding to own classes
